requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
numpy
pandas
python-dotenv